+++ src/main/java/kitchenpos/product/controller/ProductRestController.java
@@ -1,7 +1,7 @@
-package kitchenpos.controller;
+package kitchenpos.product.controller;
 
-import kitchenpos.bo.ProductBo;
-import kitchenpos.model.Product;
+import kitchenpos.product.bo.ProductBo;
+import kitchenpos.product.model.Product;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
